+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/accounting_accounts.py
@@ -29,12 +29,12 @@ class AccountingAccounts:
         
         url = utils.generate_url(operations.CreateAccountingAccountRequest, base_url, '/companies/{companyId}/connections/{connectionId}/push/accounts', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "accounting_account", 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, "accounting_account", True, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         query_params = utils.get_query_params(operations.CreateAccountingAccountRequest, request)
         headers['Accept'] = 'application/json'
-        headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
+        headers['user-agent'] = self.sdk_configuration.user_agent
         
         client = self.sdk_configuration.security_client
         
@@ -48,7 +48,7 @@ class AccountingAccounts:
 
         def do_request():
             return client.request('POST', url, params=query_params, data=data, files=form, headers=headers)
-        
+
         http_res = utils.retry(do_request, utils.Retries(retry_config, [
             '408',
             '429',
@@ -89,7 +89,7 @@ class AccountingAccounts:
         url = utils.generate_url(operations.GetAccountingAccountRequest, base_url, '/companies/{companyId}/data/accounts/{accountId}', request)
         headers = {}
         headers['Accept'] = 'application/json'
-        headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
+        headers['user-agent'] = self.sdk_configuration.user_agent
         
         client = self.sdk_configuration.security_client
         
@@ -103,7 +103,7 @@ class AccountingAccounts:
 
         def do_request():
             return client.request('GET', url, headers=headers)
-        
+
         http_res = utils.retry(do_request, utils.Retries(retry_config, [
             '408',
             '429',
@@ -143,7 +143,7 @@ class AccountingAccounts:
         headers = {}
         query_params = utils.get_query_params(operations.ListAccountingAccountsRequest, request)
         headers['Accept'] = 'application/json'
-        headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
+        headers['user-agent'] = self.sdk_configuration.user_agent
         
         client = self.sdk_configuration.security_client
         
@@ -157,7 +157,7 @@ class AccountingAccounts:
 
         def do_request():
             return client.request('GET', url, params=query_params, headers=headers)
-        
+
         http_res = utils.retry(do_request, utils.Retries(retry_config, [
             '408',
             '429',