+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/models/shared/customer.py
@@ -12,9 +12,9 @@ from typing import Optional
 
 @dataclasses.dataclass
 class Customer:
-    customer_options: Optional[list[shared_option.Option]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('customerOptions'), 'exclude': lambda f: f is None }})
+    customer_options: Optional[list[shared_option.Option]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('customerOptions') }})
     r"""List of customer options from the list of customer records on the accounting software."""
-    selected_customer_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('selectedCustomerId'), 'exclude': lambda f: f is None }})
+    selected_customer_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('selectedCustomerId') }})
     r"""Selected customer id from the list of customer records on the accounting software."""