+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/models/shared/feessupplier.py
@@ -12,9 +12,9 @@ from typing import Optional
 
 @dataclasses.dataclass
 class FeesSupplier:
-    selected_supplier_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('selectedSupplierId'), 'exclude': lambda f: f is None }})
+    selected_supplier_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('selectedSupplierId') }})
     r"""Selected supplier id from the list of supplier records on the accounting software."""
-    supplier_options: Optional[list[shared_option.Option]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('supplierOptions'), 'exclude': lambda f: f is None }})
+    supplier_options: Optional[list[shared_option.Option]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('supplierOptions') }})
     r"""List of supplier options from the list of supplier records on the accounting software."""