+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/models/shared/groupingperiod.py
@@ -11,9 +11,9 @@ from typing import Optional
 
 @dataclasses.dataclass
 class GroupingPeriod:
-    grouping_period_options: Optional[list[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('groupingPeriodOptions'), 'exclude': lambda f: f is None }})
+    grouping_period_options: Optional[list[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('groupingPeriodOptions') }})
     r"""Array of grouping period options."""
-    selected_grouping_period: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('selectedGroupingPeriod'), 'exclude': lambda f: f is None }})
+    selected_grouping_period: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('selectedGroupingPeriod') }})
     r"""Grouping period i.e. Daily sales."""