+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/models/shared/servicecharge.py
@@ -26,7 +26,7 @@ class ServiceChargeTaxComponentAllocationTaxComponentRef:
 
 @dataclasses.dataclass
 class ServiceChargeTaxComponentAllocation:
-    rate: Optional[Decimal] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('rate'), 'encoder': utils.decimalencoder(True, False), 'decoder': utils.decimaldecoder, 'exclude': lambda f: f is None }})
+    rate: Optional[Decimal] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('rate'), 'encoder': utils.decimalencoder(True, False), 'decoder': utils.decimaldecoder }})
     r"""Tax amount on order line sale as available from source commerce platform."""
     tax_component_ref: Optional[ServiceChargeTaxComponentAllocationTaxComponentRef] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('taxComponentRef'), 'exclude': lambda f: f is None }})
     r"""Taxes rates reference object depending on the rates being available on source commerce package."""